+++ src/server/game/Entities/AreaTrigger/AreaTrigger.h
@@ -1,20 +1,10 @@
-/*
- * Copyright (C) 2012-2013 JadeCore <http://www.pandashan.com/>
- * Copyright (C) 2008-2013 TrinityCore <http://www.trinitycore.org/>
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License as published by the
- * Free Software Foundation; either version 2 of the License, or (at your
- * option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
- * more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program. If not, see <http://www.gnu.org/licenses/>.
- */
+////////////////////////////////////////////////////////////////////////////////
+//
+//  MILLENIUM-STUDIO
+//  Copyright 2016 Millenium-studio SARL
+//  All Rights Reserved.
+//
+////////////////////////////////////////////////////////////////////////////////
 
 #ifndef TRINITYCORE_AREATRIGGER_H
 # define TRINITYCORE_AREATRIGGER_H
@@ -283,7 +273,7 @@ class AreaTrigger : public WorldObject, public GridObject<AreaTrigger>
 
         uint32 GetCreatedTime() const { return m_CreatedTime; }
 
-        uint32 GetUpdateInterval() const { return m_UpdateTimer.GetInterval(); }
+        uint32 GetUpdateInterval() const { return uint32(m_UpdateTimer.GetInterval()); }
         AreatriggerInterpolation GetTrajectory() const { return m_Trajectory; }
 
         AreaTriggerTemplateList GetTemplates() const { return m_Templates; }
